Beginner/requirements.txt
django==3.0.7
djangorestframework==3.11.0
pillow==7.1.2